pytorch/recommendation_system/news_recommendation/requirements.txt
nltk==3.6.5
tqdm==4.62.3
numpy==1.19.5
torch==1.9.1
scikit-learn==0.24.2
pytorch-nlp==0.5.0